added check for /proc file

+++ eventcmd.sh
@@ -5,10 +5,10 @@
 
 rec_pid=`cat "${XDG_HOME_CONFIG:-${HOME}/.config}/pianobar/record_pid"`
 if [ $? -ne 0 ]
-do
+then
 	echo "process id not found"
 	exit 2
-done
+fi
 echo record process: $rec_pid
 
 case "$1" in
@@ -16,6 +16,11 @@ case "$1" in
 
 		#arecord open file descriptor, others were NULL
 		proc_last_file=`readlink /proc/$rec_pid/fd/3`
+		if [ ! -f proc_last_file ]
+		then
+			echo could not retrieve wav file name from /proc
+			exit 2
+		fi
 		echo current file: $proc_last_file 
 		kill -s SIGUSR1 $rec_pid
 		path=$(dirname $proc_current_file)
